requirements_fixed_v2.txt
# --- LIBRERÍAS DE LA INTERFAZ GRÁFICA ---
Pillow>=10.3.0,<11
customtkinter==5.2.2
tkcalendar==1.6.1
ttkthemes==3.2.2

# --- LIBRERÍAS DE LÓGICA Y DATOS (versiones estables) ---
pandas>=2.2,<3
numpy==1.26.4
openpyxl==3.1.2

# --- LIBRERÍAS DE SEGURIDAD (versiones con pre-compilados) ---
bcrypt==4.1.2
cffi>=1.17.1
pycparser==2.22

# --- LIBRERÍAS DE REPORTES Y GRÁFICOS ---
matplotlib==3.8.4
reportlab==4.1.0

# --- OTRAS DEPENDENCIAS ---
python-dateutil==2.8.2
packaging==23.2
six==1.16.0
pytz==2024.1
tzdata==2024.1
